project1/curd_app/serilizers.py
from rest_framework import serializers
from .models import ProductOrder


class ProductOrderSerializer(serializers.ModelSerializer):
    class Meta:
        model = ProductOrder
        fields = '__all__'
